+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ### Added
 
+### Fixed
+
+### Changed
+
+## [0.2.0]
+
+### Added
+
  - CLI command for generating QC report for CDM
  - Parsing of Virulencefinder STX typing